feat(install): added --skip-npm option

+++ nove.sh
@@ -17,10 +17,18 @@ case $1 in
         ln -s /etc/node-v$2-linux-x64/bin/node /usr/bin/node
         echo -e "\e[1;2m[3/5]\e[0m \e[1;33mReplacing Node binary...\e[0m"
 
-        # Replace NPM binary
-        rm -rf /usr/bin/npm
-        ln -s /etc/node-v$2-linux-x64/bin/npm /usr/bin/npm
-        echo -e "\e[1;2m[4/5]\e[0m \e[1;33mReplacing NPM binary...\e[0m"
+        if [ "$3" = "--skip-npm" ];
+        then
+            echo -e "\e[1;2m[4/5]\e[0m \e[1;33mSkipping NPM installation...\e[0m"
+        else
+            rm -rf /usr/bin/npm
+            ln -s /etc/node-v$2-linux-x64/bin/npm /usr/bin/npm
+
+            rm -rf /usr/bin/npx
+            ln -s /etc/node-v$2-linux-x64/bin/npx /usr/bin/npx
+
+            echo -e "\e[1;2m[4/5]\e[0m \e[1;33mReplacing NPM binary...\e[0m"
+        fi
 
         # Remove Node archive
         rm -rf /etc/node-v$2-linux-x64.tar.xz
